move reset_password to /api/auth/
